add(ex_5): add ex_5 (1st ex 2nd from listing) solution, and necessary files for documentation

+++ exercises/1_Fundamentals_of_the_Java_Programming_Language/exercise_list/src/main/java/com/exercises/list_2/ex_1/ex_5.java
@@ -3,11 +3,13 @@ package com.exercises.list_2.ex_1;
 import main.java.com.exercises.list_2.GeneralMethods;
 
 public class ex_5 extends GeneralMethods {
-    
+
     private void printMultiplicationTable(int number) {
         System.out.println("Multiplication Table for " + number + ":");
-        for (int i = 1; i <= 10; i++) {
-            System.out.println(number + " x " + i + " = " + (number * i));
+
+        // Uses Long type for the loop variable to handle larger multiplication results without overflow
+        for (long i = 1; i <= 10; i++) {
+            System.out.println(number + " x " + i + " = " + ((long) number * i));
         }
     }
